Here's to Rational Approximation Week 02022!
Initial summoning phase complete.  The result appears to be a vigorous,
newborn philosophy of math education.  Warning: extreme non-linear
anomalies detected.  We've received a practical report from the unknown:

    But I was wrong.  I was somewhere new.  I became an explorer.
    An explorer of a realm inspired by the thoughts of humanity.
    I saw so many things.  And one day, I found the most extraordinary
    thing of all.  Something that could change everything.  I saw the
    means to realize our potential.

An additional practical report was received from the engineering team:

    How is this even possible?

    I have no idea.

    Do you think "Stranger Things" has actually happened?

    I seriously doubt that, but things have happened that are
    Stranger than Fiction.

This release is very late, but also hopefully right on time.