Add D6 calculators
Two mechanical calculators for computing arithmetic of the Symmetry
Group of the Hexagon

I have a variation for D4, as well as worksheets for using the
calculator, unfortunately I seem to have misplaced those pdf files.  I
should be able to find them, and if not they will be recreated.