Assorted additions
Big thanks to Jeremy Gibbons who pointed out that the average of 2 numbers
is an example of a commutative operation that is not associative, and
to Callam McGill who pointed out that the absolute difference also falls
into this category.

Also I need to revise a footnote, because now I feel like I know enough
about the relevance of the tetheral quasigroup to be comfortable using it
as an example in a relatively introductury, general-interest math course.

The tetheral quasigroup consists of a binary operation over a domain of
three distinct elements.  When the inputs to the operation differ, then
the operation returns the one that isn't among the inputs.  To make
this a quasigroup, when the operation is fed two copies of the same
element, it simply returns that element.

Quasigroups are intimately related to Latin Squares, one of the first kinds
of combinatorial designs that have been studied in the mathematical literature.

Combinatorial designs are intimately related to experimental design.

Also possibly interesting as the quasigroups of order 4 might be a
particularly simple way to introduce isotopy, as these aeem to be
an exceptionally simple nontrivial example of that.

Not as general-interest as D4, at least in the traditional math
cirriculum, but at least I have a couple answers ready to go if/when
people ask "why this topic?"