Dih(Z) is not isomorphic to its automorphism group.
Dih(Z) is centerless, so it is isomorphic to its inner automorphism group via
the conjugation map. However it has outer automorphisms. Consider
Dih(Z) = Z_2 * Z_2,  the free product of two involutions. Swapping the
involutions results in an automorphism that can be expressed as the piecewise
combination of two conjugations, but is not equivalent to any single
conjugation. Thus this automorphism is outer.

I'm pretty sure this misconception was a result of me reading this stack
overflow answer, and not doing enough sanity checking:

https://math.stackexchange.com/questions/2433055/group-isomorphic-to-its-automorphism-group

The answer specifically mentions "A Note on Groups with Just-Infinite
Automorphism Groups", by Francesco de Giovanni and Diana Imperatore, which
has this spurious claim.  It might be interesting to delve into the
support behind this claim and figure out exactly where it went wrong.

I discovered this discrepancy when I was thinking a little about automorphism
towers, and ran across this:

https://math.stackexchange.com/questions/1008972/the-automorphism-group-of-the-infinite-dihedral-group

Browsing "The Automorphism Tower Problem" by Simon Thomas I found especially
helpful. I should take a closer look at it at some point.

http://www.math.rutgers.edu/~sthomas/book.ps